Adding Java Program to Extract Last Two Digits of a Given Year
Initialization:
int x, y, i = 0;
String z = "";
Scanner s = new Scanner(System.in);
x is an integer to store the year.
y is an integer to store each extracted digit.
i is a counter initialized to 0.
z is an empty string to accumulate the last two digits.
System. out.print("Enter Year:");
x = s.nextInt();
The program prompts the user to enter a year.
Assume the user enters 2014. Therefore, x is now 2014.
First Iteration of the Loop:
while (i < 2) {
    y = x % 10; // Extract the last digit
    x = x / 10; // Remove the last digit from the number
    z = y + "" + z; // Append the extracted digit to the result string
    i++;
}
First Iteration (i = 0):
y = x % 10; calculates 2014 % 10 which is 4. So, y = 4.
x = x / 10; calculates 2014 / 10 which is 201. So, x = 201.
z = y + "" + z; updates z by concatenating 4 with the existing value of z (which is an empty string). So, z = "4".
i++ increments i to 1.
Second Iteration of the Loop:

Second Iteration (i = 1):
y = x % 10; calculates 201 % 10 which is 1. So, y = 1.
x = x / 10; calculates 201 / 10 which is 20. So, x = 20.
z = y + "" + z; updates z by concatenating 1 with the existing value of z (which is "4"). So, z = "14".
i++ increments i to 2.
Exit Loop:

The loop exits because i is now 2 which is not less than 2.
Output the Result:
System.out.println("Last two digits:" + z);
The program prints Last two digits: 14.